fix dice not rolling after resize while hidden
if you hide the canvas by using World.hide() and then resize the
browser the world is resized with width and height 0 because of
'display: none'.

fixed this by resizing the world after calling World.show().

removed the resize event listener to allow library users to handle the
resize event by them self and changed the resizeWorld function to
resize the world instead of registering the resize callback.